This is a preliminary fix to the issue identified in PR#2065
This changes softlist loading so that it performs a pass prior to setting up all slots in order to identify what softlist options need to be loaded.  In a very hacky way, I rewrite the command line to essentially expand the argc/argv to encompass softlist loading.

I don't consider this to be a final fix because rewriting the command line is gross, and I wonder if there needs to be an abstraction layer inserted where slot and image options are resolved that could either go to the command line, or resolved softlist items.

I would also argue that this approach doesn't go far enough.  If we figure out a better abstraction for slot and image setup at the beginning, all softlist handling could be hypothetically pulled out of diimage.cpp and be hidden behind that abstraction.