The recent softlist cleanup broke -console; this fixes it
The reason is that the LUA console, HTTP server and other features relied on the command line arguments being "pre-parsed".  Part of the softlist cleanup was to sanitize where command line parsing occurs.  This change moves the setup of the LUA console (and other capabilities such as the HTTP server) to the correct place in the new order.