wildpkr, tabpkr: Work out several device hookups (nw)
The use of a DS2401 "ID" in tabpkr is inferred from the code. The data supplied for it is totally artificial, and the chip may have been stripped from the PCB pictured on citylan.it.

Much still needs to be learned about the HW of these games. The emulation doesn't display even error codes yet.