Fix crashes in -listxml in debug builds * Make -listxml instantiate slot devices in slots rather than under the machine root * Use dynamic_cast rather than static_cast or downcast on parent device in TI-99 slot/card devices
(nw) There are still very real issues present.  When -listxml, -listroms, -verifyroms, etc.
instantiate individual devices, they don't have the information necessary to instantiate the
expected parent device.  So slots won't be children of the bus they expect to be attached to,
and cards won't be in their usual slots.  It's unsafe to use static_cast<> on owner at the
best of times, and doing anything with the result is likely to cause silent memory corruption
during listxml (this was previously happening with the TI-99 peripheral box slot device).
You can't use downcast<> at configuration complete time either, as that degenerates to a
static_cast<> in release builds anyway, and it will cause a crash in -listxml in debug builds
as we had before.  The choices are to use a dynamic_cast and check the result before doing
anything with it before device start time, or wait for device start time and use downcast<>
at that point.