Preliminary netlist sound for the Zaccaria 1B11142 board * Tromba (trumpet) sound is not working - requires Schmitt trigger device * Connecting cassa (bass drum) swamps other instruments so it's disconnected for now * Mixing melody sound with speech/SFX is not done in netlist (should be) * Relative levels of melody/speech/SFX are probably still wrong
(nw) A good test case for this is the Money Money driver (monymony).
There's a bit of buzzing on this one as well.  The problem with the
cassa could be caused by running into non-ideal characteristics of opams
again (the LM3900 seems to ignore the V+ value supplied to it).  When
the netlist library gets Schmitt trigger support, the tromba can be
completed.  Unfortunately, the tromba is a key part of the
characteristic sound of these boards, so you really notice when it's
lacking.