Possible fix for #2496.  See that issue for analysis.
I must admit that I am not 100% confident in this fix.  This fix transplants logic that was performed in mameopts.cpp to the UI code.  The single most challenging aspect of software lists has been understanding the intention of existing behavior, and it is very difficult to eliminate the possibility that there is some aspect of how software lists are supposed to work and as a consequence, regressions.

Another possibility is that these "four word" software options should be considered a real thing and properly handled within emu_options::evaluate_initial_softlist_options().