Fixed an options overhaul regression
This crash (discovered by Wizz) had the following symptoms:
1.  Start MAME
2.  Choose "Configure Machine"
3.  Choose "Video Options"
CRASH

This was the result of the options editor not having a fully formed list of options where it was expecting one.  The fix is to change the declaration of emu_options to one that have full OSD options (it is possible that SDLMAME needs something slightly different)

Please scrutinize this fix.  I hate it.  I also hate how all of the OSD options are duplicatively declared in submenu.cpp, but with a different data structure.  At the very least we should be having some sort of factory to create "full" options for the pertinent platform; but I'm not sure how precisely we would want it to work.