Fixed an issue that could cause a crash after performing a "Reset" after changing slot options
Steps to reproduce:
1. mame64 -window c64 -iec11 c1526
2. Open the slots menu
3. Change 'c1526' to 'fd4000'
4. Reset

The issue was that when options_extract() was ran, the changing slot options could cause the name of the option to change, so I changed this logic to use the canonical instance name.  This was similar to other issues related to the need to use canonical instance names.