Fix to MameTesters bug #6657 (#2560)
The source of the issue is that the logic for rotating slot options was broken.  It would count selectable options and skip over options that were not selectable, except for the specific scenario where the index was zero.  I ended up largely replacing the logic to be more C++-ish.