HP HIL: passthrough commands between devices
The old code didn't loopback commands that are sent to a
non-existing HIL address. Unfortunately the HP 9000/300 Boot
ROM does that: it send a IDD command to every possible address
and fails if it's not looped back.

This also changes the API. HIL devices are now able to return true
if the command should be passed on to the next devices, or false
it it shouldn't. This is required for initial configuration of the Bus
and various other commands like POLL, which are sent to the global
address.

Signed-off-by: Sven Schnelle <svens@stackframe.org>