new NOT WORKING clones Maniac Square (protected, Version 1.0, Checksum DEEE) [Peter Wilhelmsen]
came from a working protected board so we're going to attempt to dump the SRAM and verify the dump we have using it before committing that.  All going well this will be promoted to working in a few days, all going badly, we'll have to find another.  Either way, it appears to be a newer set due to extra code being inserted causing lots of other code to be offset (despite still identifying as Version 1.0) and becomes the new parent