new clones Grudge Match (V00.90, Italy, location test?) [Vernimark, f205v]
note, dump is a bit odd, 2 roms were identical, one rom needs to be loaded over the end of another as not to get corrupt text.  I've asked for verification.
despite being in Italian this is a newer, and noticably more complete version than the previous set, so I've made it the parent