Fix Batrider's audio CPU clock.
Was doing a repair today that required me to hook up a different clock source to the YM2151, so we tried to use the Z80's clock. Turns out, it's not at all 4MHz, so it was wrong. We put the 32MHz oscillator and the Z80 clock on an oscilloscope and determined the Z80 clock was divided by 6 instead of 8. So instead we divided the 68k's clock by 4 and got 4MHz, so the 68k is indeed 16MHz.