enable portname plugin by default, turn some of sjy96525's changes into .po files
note, we probably need a better way to do this so that we don't have to duplicate or create files for every clone
my original suggestion is that we require a list of sets that the replacements apply to above each list of replacements and process 'drivername.po' as well as 'parent.po' to look for replacements.  Also having an 'import from' type logic might help keep things tidier too.

if we're going to push this as the official way to do button naming for the more obvious cases then we really need it enabled by default, which is why I made that change, otherwise it's never going to gain any momentum.

of course this whole thing is up for review.