Untangle cartslot from TV Boy, replace unnecessary bankdev with ROM bank, fix save state for TV Boy.
(nw) This is a great example of why trying to make a derived class that
removes stuff from the base is a bad idea.  The machine start hook for
a2600 both registered variables for save states and mapped the
cartridge.  The tvboy drivers couldn't call this as there's no cartridge
slot present.  This meant they didn't register for save state properly
either.  It may seem tempting to just make devices optional in the base
class, but you lose useful validation and set yourself up for subtle
issues like this.  It's better to separate out the common parts into a
base class.