pc98_cd.xml: additions, testing & fixes
- Updated the CD software list with everything from the latest update of
the Neo Kobe Collection (too many new entries to list in a commit
description).
- Tested all software list entries, added supported status and emulation
issues for each one. Some games need testing on real hardware to make
sure the issues are emulation-related; I'll do this at a later date.
- Added "usage" tag for games that require serials or manual protection.
- Added disk 2 of the floppy version of Brandish Renewal, since it's
used as a key disk for the CD version.
- Fixed copy-paste error (interface="fmt_cdrom") for hybrid discs copied
from the FM Towns list.
- Replaced the Sweet Days CHD with a properly converted one.
- Some minor consistency / spelling fixes on descriptions.