mc146818: correct interrupt polarity (nw)
This change makes the MC146818/DS128x devices use the regular ASSERT_LINE/CLEAR_LINE protocol like almost all other devices, removing an unexpected surprise. The devices are fairly widely used, so I've tried to review all of those references and check if they're affected. A surprising number of systems appear to have the interrupt hooked up, but were not accounting for the inversion, so this change might possibly correct some behaviour.

The following were modified to remove workarounds:
* src/devices/machine/cs4031: removed inversion
* src/devices/machine/wd7600: removed inversion
* src/mame/drivers/apollo: hooked up interrupt
* src/mame/drivers/micronic: removed inversion
* src/mame/drivers/octopus: removed inversion
* src/mame/machine/apollo: removed timer workaround
* src/mame/machine/at: removed inversion
* src/mame/machine/interpro_ioga: removed inversion

The behaviour of the following might change:
* src/devices/bus/econet/e01.cpp
* src/devices/bus/electron/cart/click.cpp
* src/devices/bus/lpci/southbridge.cpp
* src/devices/machine/fdc37c93x.cpp
* src/devices/machine/sis85c496.cpp
* src/mame/drivers/at.cpp
* src/mame/drivers/calchase.cpp
* src/mame/drivers/hotstuff.cpp
* src/mame/drivers/hx20.cpp
* src/mame/drivers/mbee.cpp
* src/mame/drivers/mtouchxl.cpp
* src/mame/drivers/pc1512.cpp
* src/mame/drivers/pcat_dyn.cpp
* src/mame/drivers/pcd.cpp
* src/mame/drivers/qx10.cpp
* src/mame/drivers/su2000.cpp
* src/mame/machine/pcshare.cpp