(nw) add warning comment to dangerous API
I'm not sure this API is a great idea - see inline comment on https://github.com/mamedev/mame/commit/796abaf7f3200a96499440a4c40bec3440d55209

I realise it seems intuitive, but it breaks with a use-after-free (not even a
validity error) if you replace/remove the target device.  Implementing it in a
way that isn't fragile and doesn't hurt the performance of -romident,
-validate and -listxml seems impossible.  If it causes developer confusing and
things start breaking, back to literal tags we go for this case.