bt459: add contrast control (nw)
Use a simplistic multiply to simulate contrast control. InterPro systems write to a "virtual contrast DAC" register/latch of some kind, which presumably adjusts the reference voltage produced by the Bt438 before it enters the RAMDAC; this approximation is very crude, however it's simple and produces a satisfactory visual result.