new clones Four Trax (Japan) [ShouTime, The Dumping Union]
note, one of the gfx roms in the MAME set has an 'fx2' label, it should be verified.

also replaced the ROMs for the fourtraxa set, the old ones had a 0x00 fill instead of a 0xff fill but were otherwise identical, it's a weird mod of the World set that hides most of the advertising banners, replacing them with generic ones, presumably for legal reasons?