Add Sun mouse port and hook it up to sun3, sun3x, sun4, and sun4c. Also add notes to Sun keyboard emulation.
There's a hack to make the 1200 Baud mouse actually run at 9600 Baud.
This is necessary because the SCC is incorrectly expecting 9600 Baud
rather than 1200 Baud.  I don't have time to fix the SCC, so I'd
appreciate it if someone else would.

There's no way it should be expecting 9600 Baud on the mouse port.
Solaris 2.3 and later support 1200 Baud and 4800 Baud, and earlier
versions support 1200 Baud only.  No version of Solaris works with a
9600 Baud mouse.

The workaround allows the mouse to be used in SunView on sun4_60 - I
haven't tested any other drivers.