typo fixes to a couple alt_title lines
"chestfld" alt_title must have been incorrectly copied from a sales listing as the subtitle read "without box or instructions".  Replaced with correct subtitle.
Added an extra dot to "touch" subtitle to match what is printed on the instruction manual.  The box cover only shows one dot but this is due to the line break after the word "of".  In Japanese orthography the dots are certainly optional but mixing the dots and no-dots in the same phrase looks awkward.