New working clones ------------------ Street Fighter II': Champion Edition (Red Wave PtII, bootleg) [okada yuko]
(nw) This is a further hack of the sf2red program.  According to the
submitter, it was only operated in a few Japanese arcades.  He dumped it
with a ROM programmer back in the day, and recently found it on an old
hard disk.  I don't know what all the differences from sf2red are.