New WORKING clone
-----------------------
Galaxy Gunners (set 1) [David Torres, ClawGrip, Recreativas.org]

Also documented the CPUs to be 8088 (main) and V20 (sound) and corrected the sound clock (7.9955 instead of 8MHz).

Out of whatsnew: It sounds good to me, but I left MACHINE_IMPERFECT_SOUND on. I'll record the real sounds from the PCB on a few days to compare.