new NOT WORKING ClickStart [Sean Riddle, Clawgrip, TeamEurope] + a Software List for it
seems to be something missing, jumps to outside of code area, maybe secondary small internal ROM of CPU.  I did think the dumps were half size for various reasons, but there is a checksum in the footer of each one that matches the data (even if another checksum in the header for some of them we can't match)  Possible BIOS is wrong tho as that checksum in the footer doesn't match by quite a significant amount.

most games show something just by putting a 'RET' opcode in the upper areas it attempts to jump to.

Ryan credited in driver since it's more or less just a stripped down copy of vsmile.cpp