Add new softlist: original Apple II disks in WOZ format.
All existing 4AM dumps as of January 17th, 2019 are documented herein.

Tafoid assisted with a tool to automate large chunks of the process, but the tool wasn't perfect and the resulted required a second and third pass by eye.

I believe I have all the problems worked out and it passes -valid.

Last step for the moment is to go into the other Apple II softlist and start renaming the collision entries.