Fix problem with bac06 NW
The previous emulation logic change was correct, however Sly Spy, Hippodrome, Birdy Try suffered from an ordering bug - tilemap.cpp in the core cached the tile indices before the game updated the dynamic colmun/row major bit.  It seems the correct fix is to make sure tilemap::mappings_update() should be called when the column/row bit changes, however this is currently a private function.  I've put in a hack at line 396 so that mappings_update() is called as a side effect of a flip.  The better fix would be just to expose mappings_update for general driver use, but I didn't want to make that larger change without approval.