pokey: rework for performance enhancements
Profiling with valgrind pointed to the following hotspot:

pokey_device::step_one_clock / line 686-689:
| for (int ch = 0; ch < 4; ch++)
| {
| 	sum |= (((((m_channel[ch].m_output ^ m_channel[ch].m_filter_sample) || (m_channel[ch].m_AUDC & VOLUME_ONLY)) ? (m_channel[ch].m_AUDC & VOLUME_MASK) : 0 )) << (ch * 4));
| }

First solution was to move bit-shifting to first part of ?-part:

| sum |= (((m_channel[ch].m_output ^ m_channel[ch].m_filter_sample) || (m_channel[ch].m_AUDC & VOLUME_ONLY)) ? ((m_channel[ch].m_AUDC & VOLUME_MASK) << (ch * 4)) : 0);

because shifting 0 does not change value. Performance measurements showed
improvements but they were not satisfying exactly (change is part of this
patch).

So I thought more of what this piece of code is about:
* it is run at high frequency (@starwars: 1.5MHz * 4 pokey instances *
  4 channels -> ~6MHz)
  => that is creating the high CPU cycle consumption
* frequency of output change is in a range of (double) audible frequencies (few
  kHz).
  => there are long sequences creating identical output
* the sum value calculated depends on few channel input variables: m_output /
  m_filter_sample / m_AUDC

This patch suggests a solution which keeps track of possible input variable
change and as long as they don't change there is no need to render output sum.

The following tests were performed:
* mame64 -bench 50 starwars: Average speed increases from ~430 to ~490 on my PC
* on screen profiling shows ~2% idle win
* starwars, missile-command and marble-madness do not show any audible
  artefacts

Signed-off-by: Andreas Müller <schnitzeltony@gmail.com>